check if directory exists

+++ run.sh
@@ -7,13 +7,21 @@ source ./path.env
 sh get_git_info.sh
 
 # clear test cases folder
-rm -f ./jiawei-computational-results/test_cases_datatable/*
+if [[ -d ./jiawei-computational-results/test_cases_datatable ]]; then
+  rm -f ./jiawei-computational-results/test_cases_datatable/*
+else
+  mkdir ./jiawei-computational-results/test_cases_datatable/
+fi
 
 # generate test_cases
 sage ./scripts/generate_test_cases.sage
 
 # clear results folder
-rm -f ./jiawei-computational-results/results_datatable/*
+if [[ -d ./jiawei-computational-results/results_datatable ]]; then
+  rm -f ./jiawei-computational-results/results_datatable/*
+else
+  mkdir ./jiawei-computational-results/results_datatable/
+fi
 
 # submit jobs
 sbatch -t 01:00:00 ./scripts/SLURM-faster-subadditivity-test.sage
